requirements.txt
flask
flask-sqlalchemy
python-dotenv
langchain
langchain-groq
langchain-community
langchain-core
langchain-qdrant
qdrant-client
sentence-transformers
psycopg2-binary
PyPDF2
stripe
groq
python-multipart
bcrypt
werkzeug
torch
numpy
langchain_huggingface
